Rename service folder to services
